backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
pymysql==1.1.1
bcrypt==4.2.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.20
python-dotenv==1.0.1
